@@ -1,5 +1,6 @@
 import { FastifyReply, FastifyRequest } from 'fastify';
 import { ContractGeneratorService } from '../../contract-generator';
+import { ContractAnalyzerService } from '../../contract-analyzer';
 
 interface GenerateRequest {
   prompt: string;
@@ -16,8 +17,13 @@ export async function generateContractHandler(
   }
 
   const contractGenerator = new ContractGeneratorService();
+  const contractAnalyzer = new ContractAnalyzerService();
   try {
     const generatedContract = await contractGenerator.generateContract(prompt);
+    const analysis = await contractAnalyzer.analyzeContract(
+      generatedContract.source,
+      generatedContract.abi
+    );
     
     // Si hay errores de compilación, devolver un código de estado 400
     if (generatedContract.errors && generatedContract.errors.length > 0) {
@@ -35,7 +41,11 @@ export async function generateContractHandler(
       abi: generatedContract.abi,
       bytecode: generatedContract.bytecode,
       warnings: generatedContract.warnings,
-      attempts: generatedContract.attempts || 1
+      attempts: generatedContract.attempts || 1,
+      analysis: {
+        functionDescriptions: analysis.functionDescriptions,
+        diagramData: analysis.diagramData
+      }
     });
   } catch (error) {
     return reply.code(500).send({
